Move lock to handl funciton call.
Instead of locking on the stats change, we can lock on the handle
funciton of that handler so we are sure only one thread is *handling* a
message at a time. This will allow us to also ensure thread safety for
the component part.

Another solution would be to move the componenet to the parse function,
but that makes absolutely no sense to me.